infra-fs/pyproject.toml
[project]
name = "devsecops-poc-serverless"
version = "0.1.0"
description = "Serverless infrastructure for Omnichannel Publisher"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aws-cdk-lib>=2.170.0",
    "constructs>=10.3.0",
    "aws-cdk.aws-lambda-python-alpha>=2.170.0a0",
    "aws-cdk.aws-amplify-alpha>=2.170.0a0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.1.0",
]

[tool.ruff]
target-version = "py312"
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "W"]
